move item updated, unit tests for success added

+++ Necromancy.Server/Systems/Item/ItemSpawnParams.cs
@@ -7,8 +7,7 @@ namespace Necromancy.Server.Systems.Item
     public class ItemSpawnParams
     {
         public ItemStatuses ItemStatuses { get; set; }
-        public int Quantity { get; set; } = 1;
+        public byte Quantity { get; set; } = 1;
         public GemSlot[] GemSlots { get; set; } = new GemSlot[0];
-
-}
+    }
 }
